payments/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-import-export==2.7.0
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
idna==3.3
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
pycparser==2.21
PyJWT==2.3.0
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.2.0
PyYAML==6.0
reportlab==3.6.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
tzdata==2021.5
urllib3==1.26.8
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
